Backend/requirements.txt

Flask-Cors
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
requests